Remove dependency on cmake and add tmp test-build
CMake 4.0 was released, deprecating CMake < 3.5
librdkafka (the C-lib) requires a min version of 3.2 without providing a
max version. This also has been deprecated, so building with cmake
fails.
As long as librdkafka does not release with a fixed CMakeList.txt, we
need to either pin cmake-sys to a cmake version < 4.0 or try to build
with mklove.
This commit tests the latter.